+++ Flow/src/utopia/flow/view/mutable/eventful/Lockable.scala
@@ -2,7 +2,19 @@ package utopia.flow.view.mutable.eventful
 
 import utopia.flow.event.model.Destiny
 import utopia.flow.event.model.Destiny.{MaySeal, Sealed}
-import utopia.flow.view.template.eventful.MayStopChanging
+import utopia.flow.view.template.eventful.{Changing, MayStopChanging}
+
+object Lockable
+{
+	/**
+	  * Provides a lockable view into another pointer
+	  * @param other Another pointer
+	  * @tparam A Type of the changing values in the other pointer
+	  * @return A view into the other pointer which may be locked,
+	  *         after which the changes within 'other' are no longer reflected in this resulting pointer.
+	  */
+	def view[A](other: Changing[A]) = LockableBridge(other)
+}
 
 /**
   * Common trait for changing items that may be "locked".